# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Stadion Tracker API",
        "description": "API-documentatie voor stadionbezoeken, wedstrijden, posts en gebruikersinteractie.",
        "contact": {
            "email": "cedricvanhoorebeke@outlook.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://backend.ddev.site",
            "description": "DDEV ontwikkelomgeving"
        }
    ],
    "paths": {
...
        "/api/visits": {
            "get": {
                "tags": [
                    "Visits"
                ],
                "summary": "Toon alle bezoeken van de ingelogde gebruiker",
                "operationId": "59bc502e31c31bcb9d32d86e06260de5",
                "responses": {
                    "200": {
                        "description": "Lijst van bezoeken",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Visit"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Niet geauthenticeerd"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Visits"
                ],
                "summary": "Registreer een stadionbezoek",
                "operationId": "2ff60d5091a706b6942379521ae9074f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "game_id"
                                ],
                                "properties": {
                                    "game_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "visited_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-12"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Leuk stadion met goede sfeer."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bezoek geregistreerd",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Visit"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Niet geauthenticeerd"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/visits/{id}": {
            "get": {
                "tags": [
                    "Visits"
                ],
                "summary": "Toon een specifiek stadionbezoek",
                "operationId": "42d6ce6dce87180d7400f1a2ad20259e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID van het bezoek",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bezoekdetails",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Visit"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Geen toegang tot dit bezoek"
                    },
                    "404": {
                        "description": "Bezoek niet gevonden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Visits"
                ],
                "summary": "Verwijder een stadionbezoek",
                "operationId": "17961e50292ffea006c5ff414d95307b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID van het bezoek",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Bezoek verwijderd"
                    },
                    "403": {
                        "description": "Niet gemachtigd"
                    },
                    "404": {
                        "description": "Bezoek niet gevonden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Visits"
                ],
                "summary": "Update een stadionbezoek",
                "operationId": "c08d3616b43bcf1ad1e1ef0da9594169",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID van het bezoek",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "visited_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-12"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Bijgewerkte notitie"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bezoek geüpdatet",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Visit"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Niet gemachtigd"
                    },
                    "404": {
                        "description": "Bezoek niet gevonden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Game": {
                "required": [
                    "stadium_id",
                    "home_team_id",
                    "away_team_id",
                    "match_date"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "stadium_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "home_team_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "away_team_id": {
                        "type": "integer",
                        "example": 4
                    },
                    "match_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-14"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2025-05-14T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2025-05-14T10:05:00Z"
                    },
                    "stadium": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Stadium"
                            }
                        ],
                        "nullable": true
                    },
                    "home_team": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Team"
                            }
                        ],
                        "nullable": true
                    },
                    "away_team": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Team"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Stadium": {
                "required": [
                    "name",
                    "latitude",
                    "longitude"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 15
                    },
                    "name": {
                        "type": "string",
                        "example": "Allianz Arena"
                    },
                    "team_name": {
                        "type": "string",
                        "example": "Bayern Munich",
                        "nullable": true
                    },
                    "profile_image": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://your-bucket.s3.region.amazonaws.com/uploads/stadiums/profile-image/abc.jpg",
                        "nullable": true
                    },
                    "banner_image": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://your-bucket.s3.region.amazonaws.com/uploads/stadiums/banner-image/xyz.jpg",
                        "nullable": true
                    },
                    "latitude": {
                        "type": "string",
                        "example": "48.2188"
                    },
                    "longitude": {
                        "type": "string",
                        "example": "11.6247"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-01T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "Team": {
                "required": [
                    "name",
                    "league_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Manchester City"
                    },
                    "league_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "profile_image": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://your-bucket.s3.region.amazonaws.com/teams/profile-image/abc.jpg",
                        "nullable": true
                    },
                    "banner_image": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://your-bucket.s3.region.amazonaws.com/teams/banner-image/xyz.jpg",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-01T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-01T10:30:00Z"
                    }
                },
                "type": "object"
            },
...
            "Visit": {
                "required": [
                    "user_id",
                    "game"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 14
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "game": {
                        "$ref": "#/components/schemas/Game"
                    },
                    "visited_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-12",
                        "nullable": true
                    },
                    "notes": {
                        "type": "string",
                        "example": "Geweldige sfeer, veel pyro!",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2025-05-14T10:15:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "example": "2025-05-14T10:20:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}